src/api/ApiUsersUserId/requirements.txt
boto3>=1.34.0
jsonschema>=4.21.0
pyyaml>=6.0.1
botocore>=1.34.0